Fixing wonky import code
